Make deletion of generated schema from this commandlet optional, requiring a command line switch to enable it

+++ SpatialGDK/Source/SpatialGDKEditor/Private/SchemaGenerator/SpatialGDKEditorSchemaGenerator.cpp
@@ -1051,10 +1051,11 @@ void CopyWellKnownSchemaFiles(const FString& GDKSchemaCopyDir, const FString& Co
 	}
 }
 
-bool RefreshSchemaFiles(const FString& SchemaOutputPath)
+bool RefreshSchemaFiles(const FString& SchemaOutputPath, const bool bDeleteExistingSchema /* = true */, const bool bCreateDirectoryTree /* = true */)
 {
 	IPlatformFile& PlatformFile = FPlatformFileManager::Get().GetPlatformFile();
-	if (PlatformFile.DirectoryExists(*SchemaOutputPath))
+
+	if (bDeleteExistingSchema && PlatformFile.DirectoryExists(*SchemaOutputPath))
 	{
 		if (!PlatformFile.DeleteDirectoryRecursively(*SchemaOutputPath))
 		{
@@ -1065,7 +1066,7 @@ bool RefreshSchemaFiles(const FString& SchemaOutputPath)
 		}
 	}
 
-	if (!PlatformFile.CreateDirectoryTree(*SchemaOutputPath))
+	if (bCreateDirectoryTree && !PlatformFile.CreateDirectoryTree(*SchemaOutputPath))
 	{
 		UE_LOG(LogSpatialGDKSchemaGenerator, Error,
 			   TEXT("Could not create schema directory '%s'! Please make sure the parent directory is writeable."), *SchemaOutputPath);
